backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.12
httpx==0.27.2
anthropic==0.40.0
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
